Fix bug where iot and file credentials failed to get cycled promptly. Reworked stream and client shutdown flow. Other minor stability and bug fixes.

+++ src/source/Common/Version.c
@@ -48,7 +48,7 @@ STATUS getUserAgentString(PCHAR userAgentName, PCHAR customUserAgent, UINT32 len
                                customUserAgent);
     }
 
-    CHK(requiredLen > 0 && requiredLen <= len, STATUS_BUFFER_TOO_SMALL);
+    CHK(requiredLen > 0 && (UINT32) requiredLen <= len, STATUS_BUFFER_TOO_SMALL);
 
 CleanUp:
